terraform/lambda_code/src/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
slack-sdk>=3.0.0
